bugfixes/small changes. massive changes/improvements
fixed the window so you can fullscreen/resize it.

updated the name and icon so it actually shows the right things.

set up a tile map and make it easy to edit using tiled.

applied a image to a sprite and made it resizable.

applied the tilemap and display it on the screen and made it resizable